Add files via upload
This is the first project that I completed for my Google Advanced Data Analytics course.  In this fictional case study, I analyzed a synthetic TikTok dataset. Each TikTok video, represented as a row in the dataset, is mainly classified as either a 'claim' or an 'opinion'. TikTok videos are classified as 'claim' when the TikTok content creator asserts a statement as fact or presents information that can be objectively verified. On the other hand, TikTok videos that are classified as 'opinion' are those where TikTok content creators are expressing their feelings or views.  
The project objectives include:
• Building a dataframe for the TikTok dataset
• Gathering descriptive statistics
• Visualizing the TikTok data in Python
• Using pandas to manipulate the data for analysis and visualization
• Conducting exploratory data analysis (EDA) to uncover patterns, trends, and relationships'
Ultimately, the final goal is to develop a machine learning model to classify claims in TikTok videos, supporting more efficient content moderation. This project serves as the foundational step toward that objective.